pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vaspera-os"
version = "0.1.0"
description = "The Autonomous Business Operating System"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Vaspera Capital", email = "hello@vasperacapital.com" }
]
keywords = [
    "ai",
    "agents",
    "automation",
    "business",
    "mcp",
    "crewai",
    "langgraph"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Agent frameworks
    "crewai>=0.80.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",

    # MCP
    "mcp>=1.0.0",

    # API clients
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-ads>=25.0.0",
    "facebook-business>=20.0.0",
    "stripe>=10.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",

    # HTTP
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",

    # Utils
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "structlog>=24.0.0",

    # Scheduling
    "apscheduler>=3.10.0",
    "rq>=1.16.0",

    # Dashboard
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/AgriciDaniel/vaspera-os"
Documentation = "https://github.com/AgriciDaniel/vaspera-os#readme"
Repository = "https://github.com/AgriciDaniel/vaspera-os"
Issues = "https://github.com/AgriciDaniel/vaspera-os/issues"

[project.scripts]
vaspera = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
